--- test/unit/core/simd/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_subdirectories()
