--- test/unit/core/simd/detail/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_subdirectories()
-
-seqan3_test(default_simd_length_builtin_simd_test.cpp)
