+++ test/unit/utility/simd/detail/default_simd_length_builtin_simd_test.cpp
@@ -9,7 +9,7 @@
 
 #include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/utility/simd/detail/default_simd_backend.hpp>
-#include <seqan3/core/simd/detail/default_simd_length.hpp>
+#include <seqan3/utility/simd/detail/default_simd_length.hpp>
 
 #include <iostream>
 #include <type_traits>
